Separate OffHeapDiskFPSet indexer tests and add additional tests.
Part of Github issue #1112 "Warning: DiskFPSet.mergeNewEntries: xxx is already on disk with hangup following.
https://github.com/tlaplus/tlaplus/issues/1112

[Tests][TLC]
