+++ tlatools/org.lamport.tlatools/test/tlc2/tool/fp/OffHeapIndexerTest.java
@@ -37,14 +37,6 @@ import util.Assert.TLCRuntimeException;
 public class OffHeapIndexerTest {
 
 	@Test
-	public void testBitshifting() throws RemoteException {
-		final int fpBits = 1;
-		final long positions = 128L;
-		final int logPos = 8;
-		doTest(fpBits, positions, logPos, new OffHeapDiskFPSet.BitshiftingIndexer(positions, fpBits));
-	}
-
-	@Test
 	public void testRescale() throws RemoteException {
 		final int fpBits = 1;
 		final long positions = 96L;
@@ -78,14 +70,6 @@ public class OffHeapIndexerTest {
 		Assert.assertEquals(0, indexer.getIdx((0xFFFFFFFFFFFFFFFFL >>> fpBits), 1));
 		// Correctly wraps around when end of array is reached twice
 		Assert.assertEquals(0, indexer.getIdx((0xFFFFFFFFFFFFFFFFL >>> fpBits), (int)positions+1));
-	}
-
-	@Test
-	public void testBitshifting2() throws RemoteException {
-		final int fpBits = 2;
-		final long positions = 128L;
-		final int logPos = 9;
-		doTest(fpBits, positions, logPos, new OffHeapDiskFPSet.BitshiftingIndexer(positions, fpBits));
 	}
 
 	@Test
@@ -179,24 +163,6 @@ public class OffHeapIndexerTest {
 		// Correctly wraps around when end of array is reached twice
 		Assert.assertEquals(0, indexer.getIdx(maxFP, (int) positions + 1));
 	}
-
-	@Test
-	public void testShift1_268435456() throws RemoteException {
-		final int fpBits = 1;
-		final long positions = 268435456L;
-
-		final Indexer indexer = new OffHeapDiskFPSet.BitshiftingIndexer(positions, fpBits);
-
-		// indexer spreads over all positions
-		Assert.assertEquals(0, indexer.getIdx(1));
-		long maxFP = 0xFFFFFFFFFFFFFFFFL >>> fpBits;
-		Assert.assertEquals(positions - 1, indexer.getIdx(maxFP));
-
-		// Correctly wraps around when end of array is reached
-		Assert.assertEquals(0, indexer.getIdx(maxFP, 1));
-		// Correctly wraps around when end of array is reached twice
-		Assert.assertEquals(0, indexer.getIdx(maxFP, (int) positions + 1));
-	}
 	
 	@Test
 	public void testRescaleMaximum() throws RemoteException {
@@ -214,29 +180,6 @@ public class OffHeapIndexerTest {
 	}
 	
 	@Test
-	public void testBitshiftOvershoot() throws RemoteException {
-		final int fpBits = 1;
-		final long positions = 536870912L;
-		
-		final Indexer indexer = new OffHeapDiskFPSet.BitshiftingIndexer(positions, fpBits);
-		Assert.assertEquals(0, indexer.getIdx(9223371952792813846L, 5));
-	}
-	
-	private void doTest(final int fpBits, final long positions, final int logPos, final Indexer indexer) {
-		Assert.assertTrue(Double.compare(Math.pow(2, logPos - fpBits), positions) == 0);
-		
-		Assert.assertEquals(fpBits, Long.numberOfLeadingZeros((positions << (Long.SIZE - logPos)) - 1));
-		
-		for (long l = 0; l < positions; l++) {
-			final long fp = l << (Long.SIZE - logPos);
-			Assert.assertEquals(l, indexer.getIdx(fp));
-			final long fpNext = ((l+1L) << (Long.SIZE - logPos)) - 1;
-			Assert.assertEquals(l, indexer.getIdx(fpNext));
-		}
-		Assert.assertEquals(0, indexer.getIdx(positions << (Long.SIZE - logPos)));
-	}
-	
-	@Test
 	public void testOverflowErrorArithmetic() {
 		try {
 			new OffHeapDiskFPSet.Indexer(Integer.MAX_VALUE + 1L, 1);
@@ -244,14 +187,5 @@ public class OffHeapIndexerTest {
 			return;
 		}
 		Assert.fail("Creation of Indexer didn't throw an exception");
-	}
-	
-	@Test
-	public void testNoOverflowErrorBitShifting() throws RemoteException {
-		try {
-			new OffHeapDiskFPSet.BitshiftingIndexer(Integer.MAX_VALUE + 1L, 1);
-		} catch (TLCRuntimeException e) {
-			Assert.fail("Creation of BitshiftingIndexer threw an exception: " + e.getMessage());
-		}
 	}
 }
